+++ src/message/__tests__/fetchActions-test.js
@@ -15,7 +15,7 @@ import {
   tryFetch,
 } from '../fetchActions';
 import { FIRST_UNREAD_ANCHOR } from '../../anchor';
-import type { ServerMessage } from '../../api/messages/getMessages';
+import type { FetchedMessage } from '../../api/rawModelTypes';
 import { streamNarrow, HOME_NARROW, HOME_NARROW_STR, keyFromNarrow } from '../../utils/narrow';
 import { GravatarURL } from '../../utils/avatar';
 import * as eg from '../../__tests__/lib/exampleData';
@@ -242,8 +242,8 @@ describe('fetchActions', () => {
     // own transformations.
     //
     // TODO: Deduplicate this logic with similar logic in
-    // migrateMessages-test.js.
-    const serverMessage1: ServerMessage = {
+    // rawModelTypes-test.js.
+    const fetchedMessage1: FetchedMessage = {
       ...message1,
       reactions: [],
       avatar_url: null, // Null in server data will be transformed to a GravatarURL
@@ -263,7 +263,7 @@ describe('fetchActions', () => {
     describe('success', () => {
       beforeEach(() => {
         const response = {
-          messages: [serverMessage1],
+          messages: [fetchedMessage1],
           result: 'success',
         };
         // $FlowFixMe[prop-missing]: See mock in jest/globalFetch.js.
@@ -353,7 +353,8 @@ describe('fetchActions', () => {
       });
 
       test("rejects when validation-at-the-edge can't handle data, dispatches MESSAGE_FETCH_ERROR", async () => {
-        // This validation is done in `migrateMessages`.
+        // This validation is done in transformFetchedMessages in
+        // rawModelTypes.
         //
         // Simulate #4156, a real-life problem that a user at server commit
         // 0af2f9d838 ran into [1], by having `user` be missing on reactions
@@ -373,7 +374,7 @@ describe('fetchActions', () => {
           // Flow would complain at `faultyReaction` if it
           // type-checked `response`, but we should ignore it if that
           // day comes. It's badly shaped on purpose.
-          messages: [{ ...serverMessage1, reactions: [faultyReaction] }],
+          messages: [{ ...fetchedMessage1, reactions: [faultyReaction] }],
           result: 'success',
         };
         // $FlowFixMe[prop-missing]: See mock in jest/globalFetch.js.
